MOVE: Code to build events to new Events class
