+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/Action.kt
@@ -21,4 +21,3 @@ sealed class Action(val name: String) {
     object Process  : Action( "Process")
     /* ktlint-enable */
 }
-
